feature: student 3 endpoints

+++ eventPlanner/src/main/java/wedoevents/eventplanner/eventManagement/dtos/ServiceBudgetItemDTO.java
@@ -12,8 +12,8 @@ public class ServiceBudgetItemDTO {
     private UUID eventId;
 
     private UUID serviceCategoryId;
-    private double maxPrice;
+    private Double maxPrice;
 
     private UUID serviceId;
-    private int serviceVersion;
+    private Double serviceVersion;
 }
